+++ src/Contracts/MigrationAnalyzerInterface.php
@@ -2,7 +2,9 @@
 
 namespace DevSite\LaravelMigrationSearcher\Contracts;
 
+use DevSite\LaravelMigrationSearcher\DTOs\MigrationAnalysisResult;
+
 interface MigrationAnalyzerInterface
 {
-    public function analyze(string $filepath, string $type): array;
+    public function analyze(string $filepath, string $type): MigrationAnalysisResult;
 }